Extract the common code in PackageInProjectReader Make it as robust & well-logged as possible Use it in all the places (4 places)
